Update 0.4.2
- Added new functions for vehicle lootdrop per parachute option.
Option can be enabled in the a3_ims config.sqf.
So far the option spawns a vehicle on a parachute after mission succsess
with the DMS_fnc_SpawnPersistentVehicle function.
For any reason the vehicle seems to be non-presistent at this state.
This is still work in progress and realy not ready for live servers!
Please only use this option if you want to give testing feedback.

- Added new toast notification and texture for vehicle drop
notification.