Updates
Added new confguration option IMS_FortressSaticMission_Arr.  This value
has to be set to succsesfull complete the DMS Mission part without
killing all NPCs.  The value depents on the load order of static DMS
missions. If the Fortress mission is your second static mission the
value is 1, if its the only active static mission or the first the value
is 0. You can cange the load order of missions in the DMS map config
file.

Thanks to Exile Community Member SeanJSOS for ponting this out!